+++ examples/callback.rs
@@ -2,7 +2,6 @@
 extern crate c_ares;
 extern crate c_ares_resolver;
 
-use std::error::Error;
 use std::sync::mpsc;
 
 use c_ares_resolver::Resolver;
@@ -10,7 +9,7 @@ use c_ares_resolver::Resolver;
 fn print_a_results(result: &c_ares::Result<c_ares::AResults>) {
     match *result {
         Err(ref e) => {
-            println!("Query failed with error '{}'", e.description());
+            println!("Query failed with error '{}'", e);
         }
         Ok(ref a_results) => {
             println!("Successful A lookup...");